Use mk_cdisk instead of assemble_cvd
mk_cdisk is a tool to create a composite disk image. It can be used to
create other kinds of composite disks as well as os_composite.os which
was created by assemble_cvd before.

For example, if we want to pass host apexes to microdroid we can pass
them via a composite disk referencing host apex files.

Bug: 181093750
Test: running microdroid VM following updated microdroid/README.md
TEst: MicrodroidTestCase
Change-Id: Ief6468564da9eb78c6729ae64a218811919b3ae0