Increase the timeout of the logwrapper
When microdroid is running in cuttlefish (which itself is running on
GCE), it usually takes more than 2 minutes to boot microdroid. However,
the default timeout value for the shell command is set to 2 min by the
tradefed library (coincidentally!). This means that the logwrapper
command might experience a timeout before microdroid boots. Increase the
timeout of the logwrapper command.

Bug: 191613547
Test: watch th
Change-Id: I3721838662a98503152b93099ce68d4b60461cfe