Remove file size from authfs flags
authfs used to take flags like `--remote-unverified-file 5:10:40960`
where the 40969 is the file size provided by the client. The intention
is that since the file is readonly, we can set up the FUSE with the
information ready, and avoid requesting the size on every stat(2)-like
syscall.

It's just that this make it harder to use for authfs' clients. This
change remove the needs for the clients to provide the size. Instead,
authfs internally just query fd_server before starting the FUSE.

Bug: 198824883
Test: atest AuthFsHostTest ComposHostTestCases
Change-Id: I9b26b6e13bfa170d766a2428cb485d97e0423e13