Test CompOS more directly.
Rather than using "vm --daemonize ..." to test, which is a slightly
unrealistic environment and doesn't exercise much of our client-side
code, use composd where we can.

A side-effect of this is that we no longer need to disable selinux for
the tests, nor do we need the special "don't log to stderr"
configuration.

Bug: 200924405
Test: atest ComposHostTestCases
Change-Id: I2c56ef0decb521a9208843fc54166e6778388879