Use keystore2_microdroid instead.
keystore2_microdroid is a variant of keystore2 which uses libsqlite
which doesn't depend on the i18n APEX. Previously, the same keystore2
module was used across Android and microdroid. However, the use of the
new libsqlite variant caused a small regression (6ms) to the boot time -
because it couldn't make use of the original libsqlite which was already
pre-loaded by zygote.

To fix the regression, keystore2 for Android goes back to use the
original libsqlite, and keystore2_microdroid is created to use the
no-icu version of libsqlite.

Bug: 201344281
Test: measure SystemServerTiming_StartLockSettingsService-total
Change-Id: I01de5d803479b48ec407e97c193d7e093cdc0ac9